requirements.txt
django
django_rest_framework 
mysqlclient
django-cors-headers